#### What this PR does
This PR introduces an additional request validation at distributor level in order to remove all float and histogram samples containing a duplicated timestamp. Namely, for each timeseries in a request, distributor iterates all its float and histogram samples, and as soon as a sample with a timestamp that has already been visited is found, the corresponding sample is discarded from the timeseries. Once the validation is completed, the request without discarded samples is forwarded to ingesters. 

Discarded samples are tracked by the `cortex_discarded_samples_total` metrics with reason `sample_duplicate_timestamp`.

Benchmarks for the following usecases have been added (all these inputs have been applied to `distributor.validateSeries()`:
- a timeseries containing one float sample
- a timeseries containing one histogram sample
- a timeseries containing one float sample and one histogram sample
- a timeseries containing two float samples without timestamp duplications
- a timeseries containing two histogram samples without timestamp duplications
- a timeseries containing two float samples and two histogram samples without timestamp duplications
- a timeseries containing 80K float samples, all with the same timestamp, such that 40K have the same value, and 40K different values,
- a timeseries containing 80K histogram samples, all with the same timestamp, such that 40K have the same histogram, and 40K different histograms,
- a timeseries containing 80K float samples and 80K histogram samples, all with the same timestamp, such that 40K float samples and 40K histogram samples have the same value, while the remaining float and histogram samples have different values.

Comparison of `go test -bench BenchmarkDistributor_SampleDuplicateTimestamp -run=^# -count=6 -benchmem`
 of the benchmarks on `main` and the PR brench produced the following results:
```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/distributor
cpu: Apple M2
                                                                                     │ withoutDeduplication.txt │  withDeduplication.txt   │
                                                                                     │         sec/op           │     sec/op     vs base   │
one_timeseries_with_one_sample-8                                                                  28.00n ±  3%    30.51n ±   3%   +8.93% (p=0.002 n=6)
one_timeseries_with_one_histogram-8                                                               28.66n ± 16%    31.55n ±   5%        ~ (p=0.056 n=6)
one_timeseries_with_one_sample_and_one_histogram-8                                                33.07n ± 14%    35.51n ±   4%        ~ (p=0.093 n=6)
one_timeseries_with_two_samples-8                                                                 32.36n ±  1%    63.52n ±   7%  +96.29% (p=0.002 n=6)
one_timeseries_with_80_000_samples_with_duplicated_timestamps-8                                   423.4µ ± 82%    339.7µ ±  94%        ~ (p=0.132 n=6)
one_timeseries_with_80_000_histograms_with_duplicated_timestamps-8                               12.686m ± 51%    6.996m ± 192%        ~ (p=0.310 n=6)
one_timeseries_with_two_histograms-8                                                              48.98n ± 27%    71.58n ±   9%  +46.13% (p=0.002 n=6)
one_timeseries_with_two_samples_and_two_histograms-8                                              62.76n ±  8%   106.85n ±   9%  +70.25% (p=0.002 n=6)
one_timeseries_with_80_000_samples_and_80_000_histograms_with_duplicated_timestamps-8            14.402m ± 44%    6.554m ±  45%  -54.49% (p=0.004 n=6)
geomean                                                                                           1.806µ          1.854µ          +2.66%

                                                                                     │ withoutDeduplication.txt │    withDeduplication.txt    │
                                                                                     │         B/op             │     B/op       vs base      │
one_timeseries_with_one_sample-8                                                                 0.000 ±  0%       0.000 ±  0%          ~ (p=1.000 n=6) ¹
one_timeseries_with_one_histogram-8                                                              0.000 ±  0%       0.000 ±  0%          ~ (p=1.000 n=6) ¹
one_timeseries_with_one_sample_and_one_histogram-8                                               0.000 ±  0%       0.000 ±  0%          ~ (p=1.000 n=6) ¹
one_timeseries_with_two_samples-8                                                                0.000 ±  0%       0.000 ±  0%          ~ (p=1.000 n=6) ¹
one_timeseries_with_80_000_samples_with_duplicated_timestamps-8                                  38.50 ± 77%     1646.00 ±  2%  +4175.32% (p=0.002 n=6)
one_timeseries_with_80_000_histograms_with_duplicated_timestamps-8                             1.107Ki ± 51%     2.197Ki ± 54%    +98.46% (p=0.002 n=6)
one_timeseries_with_two_histograms-8                                                             0.000 ±  0%       0.000 ±  0%          ~ (p=1.000 n=6) ¹
one_timeseries_with_two_samples_and_two_histograms-8                                             0.000 ±  0%       0.000 ±  0%          ~ (p=1.000 n=6) ¹
one_timeseries_with_80_000_samples_and_80_000_histograms_with_duplicated_timestamps-8          1.267Ki ± 42%     3.569Ki ±  7%   +181.80% (p=0.002 n=6)
geomean                                                                                                      ²                    +83.78%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean
                                                                                     │ withoutDeduplication.txt │ withDeduplication.txt │
                                                                                     │        allocs/op         │  allocs/op   vs base  │
one_timeseries_with_one_sample-8                                                                 0.000 ±  0%        0.000 ±   0%  ~ (p=1.000 n=6) ¹
one_timeseries_with_one_histogram-8                                                              0.000 ±  0%        0.000 ±   0%  ~ (p=1.000 n=6) ¹
one_timeseries_with_one_sample_and_one_histogram-8                                               0.000 ±  0%        0.000 ±   0%  ~ (p=1.000 n=6) ¹
one_timeseries_with_two_samples-8                                                                0.000 ±  0%        0.000 ±   0%  ~ (p=1.000 n=6) ¹
one_timeseries_with_80_000_samples_with_duplicated_timestamps-8                                  0.000 ±  0%        6.000 ±   0%  ? (p=0.002 n=6)
one_timeseries_with_80_000_histograms_with_duplicated_timestamps-8                               11.50 ± 48%        12.00 ± 108%  ~ (p=0.567 n=6)
one_timeseries_with_two_histograms-8                                                             0.000 ±  0%        0.000 ±   0%  ~ (p=1.000 n=6) ¹
one_timeseries_with_two_samples_and_two_histograms-8                                             0.000 ±  0%        0.000 ±   0%  ~ (p=1.000 n=6) ¹
one_timeseries_with_80_000_samples_and_80_000_histograms_with_duplicated_timestamps-8            13.50 ± 41%        14.00 ±  21%  ~ (p=0.853 n=6)
geomean                                                                                                      ²                    ?               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

```


#### Which issue(s) this PR fixes or relates to

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.