@duricanikolic 2024-12-10
I have fixed most of the findings. The only missing thing are improved benchmarks, that I am going to fix tomorrow.