@bboreham 2025-01-09
Prior to this we would ignore repeated timestamps with the same value; now we report them as an error.
Was that intentional?